+++ CHANGELOG.md
@@ -4,8 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1..] - 2025--
-- Add body color to stylesheet.
+## [1.4.0] - 2025-4-14
+- Add and implement transcription status model.
+- Register transcription status model on admin page.
+- Display submission status (while still pending or processing) on home page.
+- Move transcriptions to new page.
+- Catch failures for downloading, transcribing, and diarizing.
+- Update transcription model to show current status and to fail incomplete statuses.
+- Add navbar and body color.
+- Update footer.
+- Modify updates section of readme.
+- Squash migrations between this and previous release.
 
 ## [1.3.0] - 2025-03-17
 - Use update_fields on model saves when able to.